refactor(agentic-ai): move AWS proxy-builder support out of the langchain4j package
ChatModelHttpProxySupport's AWS Apache/Netty proxy-builder methods were used
by the native Bedrock provider and by the AgentCore memory-storage client
factory, neither of which is langchain4j-specific. Move them onto the
already-shared AgenticAiHttpProxySupport instead of introducing a second
proxy-support bean.
