+++ connectors/agentic-ai/connector-agentic-ai/src/main/java/io/camunda/connector/agenticai/aiagent/chatmodel/provider/langchain4j/configuration/AgenticAiLangChain4JChatModelConfiguration.java
@@ -70,13 +70,13 @@ public class AgenticAiLangChain4JChatModelConfiguration {
   @ConditionalOnMissingBean
   public BedrockChatModelFactory langChain4JBedrockChatModelFactory(
       AgenticAiConnectorsConfigurationProperties config,
-      ChatModelHttpProxySupport chatModelHttpProxySupport,
+      AgenticAiHttpProxySupport httpProxySupport,
       ChatMessageConverter chatMessageConverter,
       ToolSpecificationConverter toolSpecificationConverter,
       JsonSchemaConverter jsonSchemaConverter) {
     return new BedrockChatModelFactory(
         config.aiagent().chatModel(),
-        chatModelHttpProxySupport,
+        httpProxySupport,
         chatMessageConverter,
         toolSpecificationConverter,
         jsonSchemaConverter);
