+++ connectors/agentic-ai/connector-agentic-ai/src/main/java/io/camunda/connector/agenticai/aiagent/memory/conversation/awsagentcore/DefaultBedrockAgentCoreClientFactory.java
@@ -6,10 +6,10 @@
  */
 package io.camunda.connector.agenticai.aiagent.memory.conversation.awsagentcore;
 
-import io.camunda.connector.agenticai.aiagent.chatmodel.provider.langchain4j.ChatModelHttpProxySupport;
 import io.camunda.connector.agenticai.aiagent.memory.conversation.awsagentcore.AwsAgentCoreConversationStore.BedrockAgentCoreClientFactory;
 import io.camunda.connector.agenticai.aiagent.model.request.MemoryStorageConfiguration.AwsAgentCoreAuthentication;
 import io.camunda.connector.agenticai.aiagent.model.request.MemoryStorageConfiguration.AwsAgentCoreMemoryStorageConfiguration;
+import io.camunda.connector.agenticai.common.AgenticAiHttpProxySupport;
 import java.net.URI;
 import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
 import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
@@ -28,9 +28,9 @@ import software.amazon.awssdk.services.bedrockagentcore.BedrockAgentCoreClientBu
  */
 public class DefaultBedrockAgentCoreClientFactory implements BedrockAgentCoreClientFactory {
 
-  private final ChatModelHttpProxySupport proxySupport;
+  private final AgenticAiHttpProxySupport proxySupport;
 
-  public DefaultBedrockAgentCoreClientFactory(ChatModelHttpProxySupport proxySupport) {
+  public DefaultBedrockAgentCoreClientFactory(AgenticAiHttpProxySupport proxySupport) {
     this.proxySupport = proxySupport;
   }
